move shelters services into own dir
