+++ apps/betterangels-backend/shelters/services/bed.py
@@ -1,0 +1,48 @@
+from typing import TYPE_CHECKING, Any, Dict
+
+from django.core.exceptions import ObjectDoesNotExist
+from django.db import transaction
+from shelters.models import Bed, Shelter
+from shelters.selectors import shelter_get
+from shelters.services.utils import (
+    _BED_M2M_FIELDS,
+    _set_m2m_from_enums,
+    _validate_subset_attributes,
+)
+
+if TYPE_CHECKING:
+    from accounts.models import User
+
+
+@transaction.atomic
+def bed_create(*, user: "User", data: Dict[str, Any]) -> Bed:
+    """Create a new Bed associated with an existing Shelter.
+
+    Validates that *user* belongs to the shelter's organization via
+    ``shelter_get`` before creating the bed.
+
+    Raises:
+        ``ObjectDoesNotExist`` when the shelter is not found or the user
+        does not belong to its organization.
+        ``django.core.exceptions.ValidationError`` on invalid data.
+    """
+    data = dict(data)
+    shelter_id = data.pop("shelter_id")
+    try:
+        shelter = shelter_get(user=user, shelter_id=shelter_id)
+    except Shelter.DoesNotExist:
+        raise ObjectDoesNotExist(f"Shelter matching ID {shelter_id} could not be found.")
+
+    m2m_data: Dict[str, Any] = {k: data.pop(k) for k in list(data) if k in _BED_M2M_FIELDS and data[k] is not None}
+
+    _validate_subset_attributes(shelter, m2m_data)
+
+    # Drop None values so model defaults apply
+    scalar_data = {k: v for k, v in data.items() if v is not None}
+
+    bed = Bed(shelter=shelter, **scalar_data)
+    bed.full_clean()
+    bed.save()
+    _set_m2m_from_enums(bed, m2m_data)
+
+    return bed
